Removed incomplete delete handler. Clean indentations.
