+++ src/components/FullContactList.js
@@ -7,7 +7,7 @@ const FullContactList = ({contacts}) => (
       <div className="col-md-11 offset-md-1 d-flex mx-auto">
         <h1>Contact List</h1>
         <Link to='/contacts/new'>
-          <button type="button" className="btn btn-primary mx-3 align-self-center">Add Contact</button>
+          <button type="button" className="btn btn-primary mx-3 align-middle">Add Contact</button>
         </Link>
       </div>
     </div>
@@ -16,10 +16,10 @@ const FullContactList = ({contacts}) => (
       <div className="col-md-11 offset-md-1 d-flex mx-auto">
         <span>
         {contacts.map(c => (
-          <h4 className="text-capitalize contact" key={c.id}>
-            <Link to={`/contacts/${c.id}`}>{c.name}</Link>
-            <a className="mx-2 edit-tag">Edit</a>
-            <a className="delete-tag">Remove</a>
+          <h4 className="text-capitalize" key={c.id}>
+            <Link className="contact" to={`/contacts/${c.id}`}>{c.name}</Link>
+            <button type="button" className="btn btn-link edit-contact ml-1">Edit</button>
+            <button type="button" className="btn btn-link remove-contact ml-1">Remove</button>
           </h4>
         ))
         }
