+++ src/index.js
@@ -1,9 +1,9 @@
-import { BrowserRouter } from 'react-router-dom'
-import React from 'react'
-import ReactDOM from 'react-dom'
+import { BrowserRouter } from 'react-router-dom';
+import React from 'react';
+import ReactDOM from 'react-dom';
 import '../node_modules/bootstrap/dist/css/bootstrap.min.css';
-import App from './components/App'
-import './index.css'
+import App from './components/App';
+import './index.css';
 
 ReactDOM.render((
   <BrowserRouter>
